requirements.txt
Django
celery
redis
requests
beautifulsoup4
pillow
djangorestframework
markdown
django-filter
django-cors-headers
channels
channels-redis
daphne
